Add support for dtb and compiling in source tree
We usually compile into an output directory $KERNEL_OUT but some kernel
sources struggle with that. Common errors are "Error: file not found".
In such cases we can now set $KERNEL_OUT=$KDIR and the build will be
performed in the source directory.

dtb images can now be built and added to the zip archives.

Add links to the kernel source in each example configuration file.

Signed-off-by: Re4son <3520622-re4son@users.noreply.gitlab.com>